Cleaning up old audio code
simplified interface autoload to register players in one function, and play audio in another function.
added default music volume too, and adjusted all to 30%
introduced a new set bus volume
moved volume change to _on_volume_changed, as a reaction to user sliding the slider, which will then set the bus volume for the respective audio bus
Updated main.gd to support the new register_player function
Updated base_item to support the new function, and moved the SFX to a constant so its easier to change if needed.