looking at recipe, adding background placeholder
So there are no recipes currently input, as the tags are all null, also there's a new strictness check, if strict is true, it will care about the order of the tags
So there's still a ton left to do with ingredients, adding to the pot, when to check the soup, and the stirring mechanic also needs to be coded into the game.
And all the item resources still need to be created
The spawner needs to be setup for all the various ingredients as well
I did manage to get the background into the scene at least. from what I can tell, there is now an item spawner, a blue area 2d the player can click, to spawn a generic item. I see items can be assigned data, and it looks like there's a spoon item and dummy data.
Seems like there is fire functionality when you add object to an area below the cauldron
It looks like any items can be put into the fire, or into the cauldron
seems like the cauldron checks for an item, then adds ingredient tags, then free and clears the players hand
currently the _check_soup() is being passed, but could be called whenever, it goes through the match_recipe(ingredients) to check for tags that match the recipe.